requirements.txt
fastapi
uvicorn
transformers
peft
sentence-transformers
torch
pydantic
accelerate
bitsandbytes
python-multipart
python-dotenv
httpx
streamlit
PyPDF2
requests
flash-attn
